fix(plugin): address review comments
- Extract sanitizeSrc() into its own module so the test can import the
  real production code instead of a duplicated mirror (resolves
  #787 review comments from gemini-code-assist about test/source
  duplication).
- Drop 'src' from the upcast view matchers for both the inline-image
  and standalone-img converters. Previously the matcher required src
  to be present, which would stop CKEditor from upcasting "healed"
  <img data-htmlarea-file-uid="..."> elements that had src stripped on
  the previous save cycle (resolves #787 review comments from Copilot).
  The src is still read and sanitised when present.
- Fix whitespace in createImageViewElement: `const attributes= {` →
  `const attributes = {` for consistency with the inline variant.
